Fix: Address static analysis defects and enhance code quality
This commit addresses several static analysis defects identified by Clang-Tidy and other tools, and improves overall code quality.

- **API Server:**
  - Fixed `cert-err33-c` defect in `src/api/server.cpp` by casting the return value of `std::remove` to `void` to silence the "ignored return value" warning.

- **Memory Tier:**
  - Fixed `performance-move-const-arg` defect in `src/memory/memory_tier.cpp` by removing `std::move` from the assignment of `pattern` to `m_patterns[id]` as it is a trivial type.

- **Memory Tier Manager:**
  - Fixed `clang-diagnostic-double-promotion` defect in `src/memory/memory_tier_manager.cpp` by explicitly casting `out_block->coherence` and `out_block->stability` to `double` before printing.

- **CUDA API:**
  - Fixed `core.StackAddressEscape` and `clang-diagnostic-return-stack-address` defects in `src/gpu/cuda_api_unified.cpp` by changing the return type of `SEP_CUDA_ErrorString` to `std::string` to return a copy of the error message, preventing a stack address from being returned.
  - Added `#include <string>` to `src/gpu/cuda_api_unified.cpp` and `#ifdef __cplusplus` guards around the `std::string` definition in `src/gpu/cuda_api_unified.h`.

- **Quantum Evolution:**
  - Fixed `clang-diagnostic-double-promotion` defect in `src/quantum/evolution.cpp` by explicitly casting `M_PI` to `float` in the phase mutation calculation.

- **API Main:**
  - Fixed `cert-err33-c` defect in `src/api_main.cpp` by casting the return value of `std::signal` to `void` to silence the "ignored return value" warning.

- **Tests:**
  - Added `cetintrin_test.cpp` to `src/tests/CMakeLists.txt` to include the new CET intrinsic test in the build.
  - Added `cetintrin_test.cpp` to test the `_incsspd` intrinsic.

- **Documentation:**
  - Updated `docs/TODO.md` to reflect the fixes and remaining defects.

+++ src/quantum/evolution.cpp
@@ -154,7 +154,7 @@ inline float deterministicNoise(uint64_t& state)
                 glm::clamp(state.coherence + (nextFloat() * 2.0f - 1.0f) * sigma, 0.0f, 1.0f);
             state.stability = glm::clamp(
                 state.stability + (nextFloat() * 2.0f - 1.0f) * sigma * 0.5f, 0.0f, 1.0f);
-            state.phase += (nextFloat() * 2.0f - 1.0f) * sigma * M_PI;  // Add phase mutation
+            state.phase += (nextFloat() * 2.0f - 1.0f) * sigma * static_cast<float>(M_PI);  // Add phase mutation
             state.entropy =
                 glm::clamp(state.entropy + (nextFloat() * 2.0f - 1.0f) * sigma * 2.0f, 0.0f, 1.0f);
             mutated.position +=
@@ -418,7 +418,7 @@ inline float deterministicNoise(uint64_t& state)
                 glm::clamp(state.coherence + (rnd() * 2.0f - 1.0f) * sigma, 0.0f, 1.0f);
             state.stability =
                 glm::clamp(state.stability + (rnd() * 2.0f - 1.0f) * sigma * 0.5f, 0.0f, 1.0f);
-            state.phase += (rnd() * 2.0f - 1.0f) * sigma * M_PI;  // Add phase mutation
+            state.phase += (rnd() * 2.0f - 1.0f) * sigma * static_cast<float>(M_PI);  // Add phase mutation
             state.entropy =
                 glm::clamp(state.entropy + (rnd() * 2.0f - 1.0f) * sigma * 2.0f, 0.0f, 1.0f);
             mutated.position +=
